// FrontEnd/Front_TFC/src/app/interfaces/evento-request-dto.ts
export interface EventoRequestDto {
    nombre: string;
    descripcion: string;
    fechaInicio: string;       
    horaInicio: string;        
    horaFin: string;           
    aforoMaximo: number;
    minimoAsistencia: number;
    precio: number;
    imagen: string;
    idTipo: number;
    zona: string;             
    idRocodromo: number;
}